Major change
1. Added function new_gen to use it as counter
2. Separated init_board and draw. The first returns a grid for all
functions to use and the second draws what the user sees (each turn).
3. Kept a function to draw living cells
=> 2 functions should draw everything (for the user to see) and one
should init the board at the beginning of the game.
IMPORTANT: Some variables still are undefined, like 'grid' when passed
as an argument for draw(). I also still need to implement
neighbors + is_inside + the game_loop